refactor js, broke out into Data, List, Comp
