+++ css/_turkey_dark.css
@@ -95,7 +95,7 @@ body.turkey.dark-mode {
   background: transparent;
 }
 
-.turkey.dark-mode .list {
+.turkey.dark-mode .panel {
   background: rgba(38, 122, 140, 0.2);
   /* border: 1px solid rgba(255, 132, 0, 0.25); */
   /* border-color: hsla(31, 100%, 50%, 0.5); */
@@ -104,13 +104,13 @@ body.turkey.dark-mode {
   padding: 1.0em 1em 0em 0.25em;
 }
 
-.turkey.dark-mode .list h4 {
+.turkey.dark-mode .panel h4 {
   color: var(--orange);
   opacity: 0.75;
   margin-bottom: 0.5em;
 }
 
-.turkey.dark-mode .list ul {
+.turkey.dark-mode .panel ul {
   color: var(--white);
   opacity: 0.85;
   line-height: 1.65;
@@ -120,7 +120,7 @@ body.turkey.dark-mode {
   padding-bottom: 1em;
 }
 
-.turkey.dark-mode .list ul li {
+.turkey.dark-mode .panel ul li {
   margin-left: -0.5em;
   opacity: .75;
   padding-left: -0.3em;
@@ -131,7 +131,7 @@ body.turkey.dark-mode {
   background: transparent;
 }
 
-.turkey.dark-mode footer.turkey_footer {
+.turkey.dark-mode footer {
   color: var(--blue-md);
   background: transparent;
   margin: 1em auto 0;
@@ -140,8 +140,8 @@ body.turkey.dark-mode {
   padding: 0em 0 1em;
 }
 
-.turkey.dark-mode .turkey_footer h4,
-.turkey.dark-mode .turkey_footer p {
+.turkey.dark-mode footer h4,
+.turkey.dark-mode footer p {
   color: white;
   opacity: 0.3;
   letter-spacing: 0.07em;
@@ -152,6 +152,6 @@ body.turkey.dark-mode {
   margin-bottom: .5em;
 }
 
-.turkey.dark-mode .turkey_footer h4 {
+.turkey.dark-mode footer h4 {
   font-weight: normal;
 }
