+++ scripts/turkey/cities/City.js
@@ -1,0 +1,7 @@
+const City = cityObject => {
+  return `
+    <li>${cityObject}</li>
+  `
+}
+
+export default City
