+++ scripts/turkey/itineraries/Itinerary.js
@@ -1,0 +1,5 @@
+const Itinerary = itineraryObject => {
+  return `<li>${itineraryObject}</li>`
+}
+
+export default Itinerary
