+++ scripts/turkey/landmarks/Landmark.js
@@ -1,0 +1,5 @@
+const Landmark = landmarkObject => {
+  return `<li>${landmarkObject}</li>`
+}
+
+export default Landmark
