+++ scripts/turkey/leaders/Leader.js
@@ -1,0 +1,5 @@
+const Leader = leaderObject => {
+  return `<li>${leaderObject}</li>`
+}
+
+export default Leader
