+++ turkey.html
@@ -35,30 +35,30 @@
 
       <div class="flexbox">
 
-        <div class="list flexItems">
+        <div class="panel flexItems">
           <h4>Major Cities</h4>
-          <ul class="cities">
+          <ul class="panel__cities">
 
           </ul>
         </div>
 
-        <div class="list flexItems">
+        <div class="panel flexItems">
           <h4>Landmarks</h4>
-          <ul class="landmarks">
+          <ul class="panel__landmarks">
             
           </ul>
         </div>
 
-        <div class="list flexItems">
+        <div class="panel flexItems">
           <h4>Heads of State</h4>
-          <ul class="leaders">
+          <ul class="panel__leaders">
             
           </ul>
         </div>
 
-        <div class="list flexItems">
+        <div class="panel flexItems">
           <h4>Itinerary</h4>
-          <ul class="itinerary">
+          <ul class="panel__itinerary">
             
           </ul>
         </div>
@@ -67,7 +67,7 @@
 
     </section>
 
-    <footer class="turkey_footer">
+    <footer>
       <h4>Drew Harper</h4>
       <p>Oakland, California</p>
       <p>Ui/Ux Designer</p>
@@ -75,7 +75,7 @@
   </main>
 
 
-  <script type="module" src="./scripts/turkey/turkey.js"></script>
+  <script type="module" src="scripts/turkey/turkey.js"></script>
   <script>
     function darkMode() {
       var element = document.body;
